+++ src/main/java/io/github/janguenter/bluemap/integrateddynamics/adapter/bluemap523/BoundedDiagnostics.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: MIT
  */
-package io.github.janguenter.bluemap.integrateddynamics.adapter.bluemap522;
+package io.github.janguenter.bluemap.integrateddynamics.adapter.bluemap523;
 
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;